# docs/architecture/06-api-contracts/openapi-spec.yaml
openapi: 3.0.3
info:
  title: KMS API
  description: |
    Knowledge Management System API for file indexing, search, and duplicate detection.

    ## Authentication

    All endpoints require authentication via:
    - **API Key**: `X-API-Key` header
    - **JWT Token**: `Authorization: Bearer <token>` header

    ## Rate Limiting

    Default rate limits:
    - Free tier: 60 requests/minute
    - Pro tier: 300 requests/minute
    - Enterprise: 1000 requests/minute

    Rate limit headers are included in all responses.
  version: 1.0.0
  contact:
    name: KMS API Support
    email: api-support@kms.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Development
  - url: https://staging-api.kms.example.com/api/v1
    description: Staging
  - url: https://api.kms.example.com/api/v1
    description: Production
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /files:
    get:
      tags: [Files]
      summary: List files
      parameters:
        - name: source_id
          in: query
          schema:
            type: string
            format: uuid
        - name: mime_type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/FileStatus'
        - name: is_duplicate
          in: query
          schema:
            type: boolean
        - name: is_junk
          in: query
          schema:
            type: boolean
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, name, size_bytes]
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Files list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilesListResponse'
  /files/{file_id}:
    get:
      tags: [Files]
      summary: Get file
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: File details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Files]
      summary: Update file metadata
      parameters:
        - $ref: '#/components/parameters/FileId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_metadata:
                  type: object
      responses:
        '200':
          description: File updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'

    delete:
      tags: [Files]
      summary: Delete file
      parameters:
        - $ref: '#/components/parameters/FileId'
        - name: permanent
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '204':
          description: File deleted
...
  /files/bulk-delete:
    post:
      tags: [Files]
      summary: Bulk delete files
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [file_ids]
              properties:
                file_ids:
                  type: array
                  items:
                    type: string
                  maxItems: 100
                permanent:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Bulk delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer
                  failed:
                    type: integer
                  errors:
                    type: array
                    items:
                      type: object
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
...
    FileId:
      name: file_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    FileStatus:
      type: string
      enum: [pending, processing, indexed, failed, skipped]
    File:
      type: object
      properties:
        id:
          type: string
        source_id:
          type: string
        name:
          type: string
        path:
          type: string
        mime_type:
          type: string
        extension:
          type: string
        size_bytes:
          type: integer
        hash_sha256:
          type: string
        status:
          $ref: '#/components/schemas/FileStatus'
        indexing:
          $ref: '#/components/schemas/IndexingInfo'
        metadata:
          type: object
        duplicate_info:
          $ref: '#/components/schemas/DuplicateInfo'
        junk_info:
          $ref: '#/components/schemas/JunkInfo'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    IndexingInfo:
      type: object
      properties:
        chunks_count:
          type: integer
        has_embeddings:
          type: boolean
        indexed_at:
          type: string
          format: date-time
    DuplicateInfo:
      type: object
      properties:
        is_duplicate:
          type: boolean
        group_id:
          type: string
    JunkInfo:
      type: object
      properties:
        is_junk:
          type: boolean
        label:
          type: string
          enum: [normal, suspicious, junk]
        confidence:
          type: number
    FilesListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/File'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        cursor:
          type: string
        next_cursor:
          type: string
        has_more:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
            request_id:
              type: string
            timestamp:
              type: string
              format: date-time